Allow to reorder AS service categories.

+++ app/database/migrations/as/2014_08_20_135351_create_as_service_categories_table.php
@@ -19,7 +19,7 @@ class CreateAsServiceCategoriesTable extends Migration {
             $table->string('name');
             $table->string('description');
             $table->boolean('is_show_front');
-            $table->tinyInteger('order');
+            $table->integer('order');
             $table->foreign('user_id')
                 ->references('id')
                 ->on('users')
